feat: split agent-fromai-contract out of feel-function-contracts, error severity
Deterministic fromAi() breaks (wrong key type, missing/misplaced
toolCall. prefix, multi-segment key, duplicate key, missing key,
function-name casing, wrong context, non-entry-element use) have no
legitimate reading - the connector always silently resolves to
nothing. Reported via a new agent-fromai-contract rule at error
severity, since these can't be mid-edit states (the rule only
evaluates complete, syntactically valid fromAi() calls).

feel-function-contracts keeps the judgment-call cases (description
quality, conditional key) at warn severity, and now silently defers
structural violations to agent-fromai-contract instead of
double-reporting them.
