+++ test/config/configs.spec.js
@@ -530,6 +530,7 @@ describe('configs', function() {
     'ad-hoc-sub-process': [ 'error', { version: '8.8' } ],
     'agent-tool-documentation': [ 'warn', { version: '8.8' } ],
     'agent-tool-output-key': [ 'warn', { version: '8.8' } ],
+    'agent-fromai-contract': [ 'error', { version: '8.8' } ],
     'called-element': [ 'error', { version: '8.8' } ],
     'connector-properties': [ 'warn', { version: '8.8' } ],
     'duplicate-execution-listeners': [ 'error', { version: '8.8' } ],
@@ -580,6 +581,7 @@ describe('configs', function() {
     'ad-hoc-sub-process': [ 'error', { version: '8.9' } ],
     'agent-tool-documentation': [ 'warn', { version: '8.9' } ],
     'agent-tool-output-key': [ 'warn', { version: '8.9' } ],
+    'agent-fromai-contract': [ 'error', { version: '8.9' } ],
     'called-element': [ 'error', { version: '8.9' } ],
     'connector-properties': [ 'warn', { version: '8.9' } ],
     'duplicate-execution-listeners': [ 'error', { version: '8.9' } ],
@@ -630,6 +632,7 @@ describe('configs', function() {
     'ad-hoc-sub-process': [ 'error', { version: '8.10' } ],
     'agent-tool-documentation': [ 'warn', { version: '8.10' } ],
     'agent-tool-output-key': [ 'warn', { version: '8.10' } ],
+    'agent-fromai-contract': [ 'error', { version: '8.10' } ],
     'before-all-execution-listener': [ 'error', { version: '8.10' } ],
     'cancel-execution-listener': [ 'error', { version: '8.10' } ],
     'called-element': [ 'error', { version: '8.10' } ],
@@ -709,6 +712,7 @@ describe('configs', function() {
     'ad-hoc-sub-process': 'error',
     'agent-tool-documentation': 'warn',
     'agent-tool-output-key': 'warn',
+    'agent-fromai-contract': 'error',
     'before-all-execution-listener': 'error',
     'cancel-execution-listener': 'error',
     'element-type': 'error',
